+++ changelog.md
@@ -10,6 +10,7 @@ If you are reading this on the website, then these changes are live here.
 - [`windows` `◫`](https://uiua.org/docs/windows) can now take negative window sizes
 - Add the [`rectify` `⌅`](https://uiua.org/docs/rectify) function, which sets a function's inverse to itself
 - Add the [`setinv`](https://uiua.org/docs/setinv) modifier, which sets the inverse of a function
+- Add the [`setunder`](https://uiua.org/docs/setunder) modifier, which sets the [`under` `⍜`](https://uiua.org/docs/under)-compatible inverse of a function
 - Allow custom modifiers to use switch function syntax
 - [`sign` `±`](https://uiua.org/docs/sign) now normalizes complex numbers
 ### Interpreter